pyproject.toml
[tool.poetry]
name = "mltemplate"
version = "0.1.0"
description = ""
authors = ["pysan3 <pysan3@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
mlt = "mltemplate.trigger:main"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.3"
torch = "^2.0.1"
torchvision = "^0.15.2"
matplotlib = "^3.7.1"
tensorboard = "^2.13.0"
opencv-python = "^4.7.0.72"
opencv-contrib-python = "^4.7.0.72"
rich = "^13.3.5"
hydra-core = "^1.3.2"
pytest = "^7.3.1"
pytest-watch = "^4.2.0"
torchinfo = "^1.8.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
isort = { extras = ["pyproject"], version = "^5.12.0" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = '''
(
    migrations
    | .mypy_cache
    | .pytest_cache
    | .tox
    | venv
)
'''

[tool.autopep8]
max_line_length = 120
ignore = "F405"
aggressive = 2

[tool.isort]
include_trailing_comma = true
line_length = 120
multi_line_output = 3
